Constructing a solution
Sometimes we are asked both to find the value of an optimal solution and also to give an example how such a solution can be constructed. In the coin problem, this means that the algorithm should show how to select the coins that produce the sum x using as few coins as possible.